Implemented 0x40 subcommand60 RunSetPos

+++ src/Booma.Proxy.Packets.BlockServer/Payloads/Server/Subpayloads/Command60/UnknownSubCommand60ServerPayload.cs
@@ -29,7 +29,7 @@ namespace Booma.Proxy
 		/// <inheritdoc />
 		public override string ToString()
 		{
-			if(Enum.IsDefined(typeof(SubCommand60OperationCode), OperationCode))
+			if(Enum.IsDefined(typeof(SubCommand60OperationCode), (byte)OperationCode))
 				return $"Unknown SubCommand60: {OperationCode:X} Name: {((SubCommand60OperationCode)OperationCode).ToString()} Type: {base.ToString()} Size: {4 + UnknownBytes.Length + 2 + 2}";
 			else
 				return $"Unknown SubCommand60: {OperationCode:X} Type: {base.ToString()} Size: {4 + UnknownBytes.Length + 2 + 2}";
